# PetStore.Web/src/PetStore.Web/Controllers/UserApi.cs
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about Swagger at [http://swagger.io](http://swagger.io). In the third iteration of the pet store, we've switched to the design first approach! You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.  Some useful links: - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore) - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
 *
 * The version of the OpenAPI document: 1.0.17
 * Contact: apiteam@swagger.io
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using PetStore.Web.Attributes;
using PetStore.Web.Models;

namespace PetStore.Web.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class UserApiController : ControllerBase
    { 
        /// <summary>
        /// Create user
        /// </summary>
        /// <remarks>This can only be done by the logged in user.</remarks>
        /// <param name="user">Created user object</param>
        /// <response code="0">successful operation</response>
        [HttpPost]
        [Route("/v3/user")]
        [Consumes("application/json", "application/xml", "application/x-www-form-urlencoded")]
        [ValidateModelState]
        [SwaggerOperation("CreateUser")]
        [SwaggerResponse(statusCode: 0, type: typeof(User), description: "successful operation")]
        public virtual IActionResult CreateUser([FromBody]User user)
        {

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(User));
            string exampleJson = null;
            exampleJson = "{\r\n  \"firstName\" : \"John\",\r\n  \"lastName\" : \"James\",\r\n  \"password\" : \"12345\",\r\n  \"userStatus\" : 1,\r\n  \"phone\" : \"12345\",\r\n  \"id\" : 10,\r\n  \"email\" : \"john@email.com\",\r\n  \"username\" : \"theUser\"\r\n}";
            exampleJson = "<user>\n  <id>10</id>\n  <username>theUser</username>\n  <firstName>John</firstName>\n  <lastName>James</lastName>\n  <email>john@email.com</email>\n  <password>12345</password>\n  <phone>12345</phone>\n  <userStatus>1</userStatus>\n</user>";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default(User);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Creates list of users with given input array
        /// </summary>
        /// <remarks>Creates list of users with given input array</remarks>
        /// <param name="user"></param>
        /// <response code="200">Successful operation</response>
        /// <response code="0">successful operation</response>
        [HttpPost]
        [Route("/v3/user/createWithList")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("CreateUsersWithListInput")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "Successful operation")]
        public virtual IActionResult CreateUsersWithListInput([FromBody]List<User> user)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(User));
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);
            string exampleJson = null;
            exampleJson = "{\r\n  \"firstName\" : \"John\",\r\n  \"lastName\" : \"James\",\r\n  \"password\" : \"12345\",\r\n  \"userStatus\" : 1,\r\n  \"phone\" : \"12345\",\r\n  \"id\" : 10,\r\n  \"email\" : \"john@email.com\",\r\n  \"username\" : \"theUser\"\r\n}";
            exampleJson = "<user>\n  <id>10</id>\n  <username>theUser</username>\n  <firstName>John</firstName>\n  <lastName>James</lastName>\n  <email>john@email.com</email>\n  <password>12345</password>\n  <phone>12345</phone>\n  <userStatus>1</userStatus>\n</user>";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default(User);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Delete user
        /// </summary>
        /// <remarks>This can only be done by the logged in user.</remarks>
        /// <param name="username">The name that needs to be deleted</param>
        /// <response code="400">Invalid username supplied</response>
        /// <response code="404">User not found</response>
        [HttpDelete]
        [Route("/v3/user/{username}")]
        [ValidateModelState]
        [SwaggerOperation("DeleteUser")]
        public virtual IActionResult DeleteUser([FromRoute (Name = "username")][Required]string username)
        {

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get user by user name
        /// </summary>
        /// <param name="username">The name that needs to be fetched. Use user1 for testing. </param>
        /// <response code="200">successful operation</response>
        /// <response code="400">Invalid username supplied</response>
        /// <response code="404">User not found</response>
        [HttpGet]
        [Route("/v3/user/{username}")]
        [ValidateModelState]
        [SwaggerOperation("GetUserByName")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "successful operation")]
        public virtual IActionResult GetUserByName([FromRoute (Name = "username")][Required]string username)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(User));
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\r\n  \"firstName\" : \"John\",\r\n  \"lastName\" : \"James\",\r\n  \"password\" : \"12345\",\r\n  \"userStatus\" : 1,\r\n  \"phone\" : \"12345\",\r\n  \"id\" : 10,\r\n  \"email\" : \"john@email.com\",\r\n  \"username\" : \"theUser\"\r\n}";
            exampleJson = "<user>\n  <id>10</id>\n  <username>theUser</username>\n  <firstName>John</firstName>\n  <lastName>James</lastName>\n  <email>john@email.com</email>\n  <password>12345</password>\n  <phone>12345</phone>\n  <userStatus>1</userStatus>\n</user>";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default(User);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Logs user into the system
        /// </summary>
        /// <param name="username">The user name for login</param>
        /// <param name="password">The password for login in clear text</param>
        /// <response code="200">successful operation</response>
        /// <response code="400">Invalid username/password supplied</response>
        [HttpGet]
        [Route("/v3/user/login")]
        [ValidateModelState]
        [SwaggerOperation("LoginUser")]
        [SwaggerResponse(statusCode: 200, type: typeof(string), description: "successful operation")]
        public virtual IActionResult LoginUser([FromQuery (Name = "username")]string username, [FromQuery (Name = "password")]string password)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(string));
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);
            string exampleJson = null;
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<string>(exampleJson)
            : default(string);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Logs out current logged in user session
        /// </summary>
        /// <response code="0">successful operation</response>
        [HttpGet]
        [Route("/v3/user/logout")]
        [ValidateModelState]
        [SwaggerOperation("LogoutUser")]
        public virtual IActionResult LogoutUser()
        {

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Update user
        /// </summary>
        /// <remarks>This can only be done by the logged in user.</remarks>
        /// <param name="username">name that need to be deleted</param>
        /// <param name="user">Update an existent user in the store</param>
        /// <response code="0">successful operation</response>
        [HttpPut]
        [Route("/v3/user/{username}")]
        [Consumes("application/json", "application/xml", "application/x-www-form-urlencoded")]
        [ValidateModelState]
        [SwaggerOperation("UpdateUser")]
        public virtual IActionResult UpdateUser([FromRoute (Name = "username")][Required]string username, [FromBody]User user)
        {

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);

            throw new NotImplementedException();
        }
    }
}