# PetStore.Web/src/PetStore.Web/wwwroot/openapi-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "contact" : {
      "email" : "apiteam@swagger.io"
    },
    "description" : "This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about\nSwagger at [http://swagger.io](http://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!\nYou can now help us improve the API whether it's by making changes to the definition itself or to the code.\nThat way, with time, we can improve the API in general, and expose some of the new features in OAS3.\n\nSome useful links:\n- [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)\n- [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService" : "http://swagger.io/terms/",
    "title" : "Swagger Petstore - OpenAPI 3.0",
    "version" : "1.0.17"
  },
...
  "servers" : [ {
    "url" : "/v3"
  } ],
...
  "paths" : {
...
    "/user" : {
      "post" : {
        "description" : "This can only be done by the logged in user.",
        "operationId" : "createUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "description" : "Created user object"
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "successful operation"
          }
        },
        "summary" : "Create user",
        "tags" : [ "user" ]
      }
    },
    "/user/createWithList" : {
      "post" : {
        "description" : "Creates list of users with given input array",
        "operationId" : "createUsersWithListInput",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/User"
                },
                "type" : "array"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "Successful operation"
          },
          "default" : {
            "description" : "successful operation"
          }
        },
        "summary" : "Creates list of users with given input array",
        "tags" : [ "user" ],
        "x-swagger-router-controller" : "UserController"
      }
    },
    "/user/login" : {
      "get" : {
        "description" : "",
        "operationId" : "loginUser",
        "parameters" : [ {
          "description" : "The user name for login",
          "explode" : true,
          "in" : "query",
          "name" : "username",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "The password for login in clear text",
          "explode" : true,
          "in" : "query",
          "name" : "password",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/xml" : {
                "schema" : {
                  "type" : "string"
                }
              },
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "successful operation",
            "headers" : {
              "X-Rate-Limit" : {
                "description" : "calls per hour allowed by the user",
                "explode" : false,
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                },
                "style" : "simple"
              },
              "X-Expires-After" : {
                "description" : "date in UTC when token expires",
                "explode" : false,
                "schema" : {
                  "format" : "date-time",
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "description" : "Invalid username/password supplied"
          }
        },
        "summary" : "Logs user into the system",
        "tags" : [ "user" ]
      }
    },
...
    "/user/{username}" : {
      "delete" : {
        "description" : "This can only be done by the logged in user.",
        "operationId" : "deleteUser",
        "parameters" : [ {
          "description" : "The name that needs to be deleted",
          "explode" : false,
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid username supplied"
          },
          "404" : {
            "description" : "User not found"
          }
        },
        "summary" : "Delete user",
        "tags" : [ "user" ]
      },
      "get" : {
        "description" : "",
        "operationId" : "getUserByName",
        "parameters" : [ {
          "description" : "The name that needs to be fetched. Use user1 for testing. ",
          "explode" : false,
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Invalid username supplied"
          },
          "404" : {
            "description" : "User not found"
          }
        },
        "summary" : "Get user by user name",
        "tags" : [ "user" ]
      },
      "put" : {
        "description" : "This can only be done by the logged in user.",
        "operationId" : "updateUser",
        "parameters" : [ {
          "description" : "name that need to be deleted",
          "explode" : false,
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "description" : "Update an existent user in the store"
        },
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        },
        "summary" : "Update user",
        "tags" : [ "user" ],
        "x-swagger-router-controller" : "UserController"
      }
    }
  },
  "components" : {
...
    "schemas" : {
...
      "User" : {
        "example" : {
          "firstName" : "John",
          "lastName" : "James",
          "password" : "12345",
          "userStatus" : 1,
          "phone" : "12345",
          "id" : 10,
          "email" : "john@email.com",
          "username" : "theUser"
        },
        "properties" : {
          "id" : {
            "example" : 10,
            "format" : "int64",
            "type" : "integer"
          },
          "username" : {
            "example" : "theUser",
            "type" : "string"
          },
          "firstName" : {
            "example" : "John",
            "type" : "string"
          },
          "lastName" : {
            "example" : "James",
            "type" : "string"
          },
          "email" : {
            "example" : "john@email.com",
            "type" : "string"
          },
          "password" : {
            "example" : "12345",
            "type" : "string"
          },
          "phone" : {
            "example" : "12345",
            "type" : "string"
          },
          "userStatus" : {
            "description" : "User Status",
            "example" : 1,
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object",
        "xml" : {
          "name" : "user"
        },
        "x-swagger-router-model" : "io.swagger.petstore.model.User"
      },
...
    },
    "securitySchemes" : {
      "petstore_auth" : {
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://petstore.swagger.io/oauth/authorize",
            "scopes" : {
              "write:pets" : "modify pets in your account",
              "read:pets" : "read your pets"
            }
          }
        },
        "type" : "oauth2"
      },
      "api_key" : {
        "in" : "header",
        "name" : "api_key",
        "type" : "apiKey"
      }
    }
  }
}